Convert history to use standard history /w gravity. Zero elo, but this gives an architecture that can scale for other enhancements.

+++ Pedantic.Chess/Interfaces.cs
@@ -2,6 +2,6 @@
 {
     public interface IHistory
     {
-        public short this[Piece piece, int to] { get; }
+        public int this[Piece piece, int to] { get; }
     }
 }
